+++ native_raycasting/src/lua_raycasting.cpp
@@ -157,9 +157,9 @@ static int CameraSetMaxDistanceLua(lua_State* L){
 //endregion
 
 static int ColorHSVToRGBLua(lua_State* L){
-    int h = luaL_checknumber(L,1);
-    int s = luaL_checknumber(L,2);
-    int v = luaL_checknumber(L,3);
+    float h = luaL_checknumber(L,1);
+    float s = luaL_checknumber(L,2);
+    float v = luaL_checknumber(L,3);
     int rgb = HSVToRGBInt(h,s,v);
     lua_pushnumber(L,rgb);
 	return 1;
@@ -169,6 +169,7 @@ static int ColorRGBToHSVLua(lua_State* L){
 	int rgb = luaL_checknumber(L,1);
 	float h,s,v;
 	RGBIntToHSV(rgb,h,s,v);
+	//h = ceil(h);
     lua_pushnumber(L,h);
     lua_pushnumber(L,s);
     lua_pushnumber(L,v);
@@ -177,7 +178,7 @@ static int ColorRGBToHSVLua(lua_State* L){
 
 // Functions exposed to Lua
 static const luaL_reg Module_methods[] =
-{	
+{
 	{"camera_update", CameraUpdateLua},
 	{"camera_set_fov", CameraSetFovLua},
 	{"camera_set_rays", CameraSetRaysLua},
